implement PlatformScheduler
