+++ core/jvm/src/main/scala/zio/scheduler/PlatformSpecific.scala
@@ -18,25 +18,19 @@ package zio.scheduler
 
 import java.util.concurrent._
 
-import zio.ZIO
 import zio.duration.Duration
-import zio.internal.NamedThreadFactory
+import zio.internal.{ NamedThreadFactory, PlatformScheduler }
 
 private[scheduler] trait PlatformSpecific {
-  private[scheduler] val defaultScheduler = new Scheduler.Service {
-    import Scheduler.CancelToken
+  import PlatformScheduler.CancelToken
+
+  private[scheduler] val platformScheduler = new PlatformScheduler {
 
     private[this] val service = Executors.newScheduledThreadPool(1, new NamedThreadFactory("zio-timer", true))
 
     private[this] val ConstFalse = () => false
 
-    override def schedule[R, E, A](task: ZIO[R, E, A], duration: Duration): ZIO[R, E, A] =
-      ZIO.effectAsyncInterrupt { cb =>
-        val canceler = _schedule(() => cb(task), duration)
-        Left(ZIO.effectTotal(canceler()))
-      }
-
-    private[this] def _schedule(task: Runnable, duration: Duration): CancelToken = duration match {
+    def schedule(task: Runnable, duration: Duration): CancelToken = duration match {
       case Duration.Infinity => ConstFalse
       case Duration.Zero =>
         task.run()
